requirements.txt
alembic==1.9.4
aniso8601==9.0.1
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.3
opencv-python==4.5.4.60
packaging==23.1
Pillow==9.5.0
pip==22.3.1
protobuf==3.20.3
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==0.21.1
pytz==2022.7.1
setuptools==65.5.0
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
numpy==1.25.1
opencv-python==4.8.0.74
Pillow==10.0.0
pip==23.2.1
protobuf==4.21.12
setuptools==68.0.0
Werkzeug==2.3.6
wheel==0.41.0